Move next pages utils to root utils dir
