+++ src/pages/v2/p/[projectId]/venft/index.page.tsx
@@ -6,7 +6,7 @@ import { GetServerSideProps } from 'next'
 import { TransactionProvider } from 'providers/TransactionProvider'
 import { V2V3ProjectPageProvider } from 'providers/v2v3/V2V3ProjectPageProvider'
 import { VeNftProvider } from 'providers/v2v3/VeNftProvider'
-import { getProjectProps, ProjectPageProps } from '../utils/props'
+import { getProjectProps, ProjectPageProps } from 'utils/server/pages/props'
 
 export const getServerSideProps: GetServerSideProps<
   ProjectPageProps
