+++ src/utils/server/pages/props.ts
@@ -1,5 +1,4 @@
 import { CV_V2, CV_V3 } from 'constants/cv'
-import { FEATURE_FLAGS } from 'constants/featureFlags'
 import { JUICEBOX_MONEY_PROJECT_METADATA_DOMAIN } from 'constants/metadataDomain'
 import { readNetwork } from 'constants/networks'
 import { readProvider } from 'constants/readProvider'
@@ -8,8 +7,8 @@ import { NetworkName } from 'models/network-name'
 import { ProjectMetadataV5 } from 'models/project-metadata'
 import { V2V3ContractName } from 'models/v2v3/contracts'
 import { GetServerSidePropsResult } from 'next'
-import { featureFlagEnabled } from 'utils/featureFlags'
 import { findProjectMetadata } from 'utils/server'
+import { isV3Project } from 'utils/v2v3/cv'
 import { loadV2V3Contract } from 'utils/v2v3/loadV2V3Contract'
 
 export interface ProjectPageProps {
@@ -34,57 +33,6 @@ async function loadJBProjects() {
   )
 
   return contract
-}
-
-async function loadV3JBDirectory() {
-  const network = readNetwork.name
-  const contract = await loadV2V3Contract(
-    V2V3ContractName.JBDirectory,
-    network,
-    readProvider,
-    CV_V3,
-  )
-
-  return contract
-}
-
-async function loadV3JBController() {
-  const network = readNetwork.name
-  const contract = await loadV2V3Contract(
-    V2V3ContractName.JBController,
-    network,
-    readProvider,
-    CV_V3,
-  )
-
-  return contract
-}
-
-/**
- * Return whether a project is a v2 or v3 project.
- *
- * A project is considered a v3 project the [projectId]
- * in the V3 JBDirectory contract has its controller set to the V3 JBController address.
- */
-async function isV3Project(projectId: number): Promise<boolean> {
-  if (!featureFlagEnabled(FEATURE_FLAGS.V3)) {
-    return Promise.resolve(false)
-  }
-
-  const [V3JBDirectory, V3JBController] = await Promise.all([
-    loadV3JBDirectory(),
-    loadV3JBController(),
-  ])
-  if (!V3JBDirectory) {
-    throw new Error(`contract not found ${V2V3ContractName.JBDirectory}`)
-  }
-  if (!V3JBController) {
-    throw new Error(`contract not found ${V2V3ContractName.JBController}`)
-  }
-
-  const projectControllerAddress = await V3JBDirectory.controllerOf(projectId)
-
-  return projectControllerAddress === V3JBController.address
 }
 
 async function getMetadataCidFromContract(projectId: number) {
